Fixed an incompatibility issue with WordPress 5.5 regarding input radio buttons.
In WordPress 5.5, selection a radio button causes all unchecked in a radio group although the attribute still holds the `checked` state. This seems to be due to the update of jQuery in the WordPress core.

This change includes to be using `.prop( 'checked', true )` instead of `.attr( 'checked', true )`, which seems to be the culprit.